Manipulation I/O Control
Updated things under the hood so that manipulations can take multiple inputs and write their output to a different slot than the input image.

Overall, all these new changes tend to revolve around the new ImageContainer.pilImagesTemp list

Same is be possible for scaling

NOTE: I feel that in this commit my code got significantly disorganized; I ran into a lot of issues breaking several functionalities while restructuring. I think this is a sign that I need to do further abstracting; It might be time to do something in terms of moving more content to StateContainer or creating a ButtonContainer class (which could even be extended to RadiobuttonContainer/CheckbuttonContainer etc.)